% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{as_proc_transform}
\alias{as_proc_transform}
\title{Expand a string shortcut into a transform specification.}
\usage{
as_proc_transform(transform)
}
\arguments{
\item{transform}{A \code{proc_transform_spec} or a string such as \code{"similarity"}.}
}
\value{
A \code{proc_transform_spec}.
}
\description{
Expand a string shortcut into a transform specification.
}
