% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aligned.R
\name{block_apply}
\alias{block_apply}
\title{Apply a function to a materialized aligned view.}
\usage{
block_apply(x, FUN, ...)
}
\arguments{
\item{x}{A \code{proc_aligned_view}.}

\item{FUN}{A function of a matrix.}

\item{...}{Passed to \code{FUN}.}
}
\description{
Apply a function to a materialized aligned view.
}
