% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{canonicalize}
\alias{canonicalize}
\title{Canonical presentation of a fit (principal axes of the consensus).}
\usage{
canonicalize(object, ...)
}
\arguments{
\item{object}{A \code{gpa_fit}.}

\item{...}{Unused.}
}
\description{
This is a presentation, not part of estimation.
}
