% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{compose_proc_transform}
\alias{compose_proc_transform}
\title{Compose two fitted transforms: \code{tr2} after \code{tr1}.}
\usage{
compose_proc_transform(tr1, tr2)
}
\arguments{
\item{tr1, tr2}{Fitted transforms.}
}
\description{
Compose two fitted transforms: \code{tr2} after \code{tr1}.
}
