% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{decompose}
\alias{decompose}
\title{Energy decomposition of aligned configurations.}
\usage{
decompose(object, ...)
}
\arguments{
\item{object}{A \code{gpa_fit}, or a list of aligned matrices when \code{consensus} is supplied.}

\item{...}{Unused.}

\item{consensus}{Optional consensus matrix.}

\item{weights}{Configuration weights.}
}
\description{
Energy decomposition of aligned configurations.
}
