% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{degrees_of_freedom}
\alias{degrees_of_freedom}
\title{Degrees of freedom of a transform family in dimension \code{d}.}
\usage{
degrees_of_freedom(spec, d)
}
\arguments{
\item{spec}{Transform specification.}

\item{d}{Dimension.}
}
\description{
Degrees of freedom of a transform family in dimension \code{d}.
}
