% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{diagnose}
\alias{diagnose}
\title{Rank, overlap, and stationarity diagnostics.}
\usage{
diagnose(object, ...)
}
\arguments{
\item{object}{A fit.}

\item{...}{Unused.}
}
\description{
Rank, overlap, and stationarity diagnostics.
}
