% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{explain_solver}
\alias{explain_solver}
\title{Explain the selected solver for a compiled problem or fit.}
\usage{
explain_solver(x)
}
\arguments{
\item{x}{A compiled problem or \code{gpa_fit}.}
}
\description{
Explain the selected solver for a compiled problem or fit.
}
