% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprocrustes-package.R
\docType{package}
\name{gprocrustes-package}
\alias{gprocrustes}
\alias{gprocrustes-package}
\title{gprocrustes: a solver-transparent Procrustes engine}
\description{
Pairwise and generalized Procrustes analysis with explicit transformation
groups, gauges, missingness, and optimality claims.
}
\details{
The mathematics in \code{docs/spec/00-mathematics.md} is the primary specification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/gprocrustes}
  \item Report bugs at \url{https://github.com/bbuchsbaum/gprocrustes/issues}
}

}
\author{
\strong{Maintainer}: Brad Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
