% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{horizontal_component}
\alias{horizontal_component}
\title{Horizontal part of a consensus perturbation (math sections 25-26).}
\usage{
horizontal_component(Z, M, spec = proc_orthogonal("O"))
}
\arguments{
\item{Z}{Perturbation matrix.}

\item{M}{Consensus.}

\item{spec}{Transform family that defines the orbit.}
}
\value{
A matrix the same shape as \code{Z}.
}
\description{
Horizontal part of a consensus perturbation (math sections 25-26).
}
\examples{
M <- scale(matrix(rnorm(20), 10, 2), scale = FALSE)
Z <- matrix(rnorm(20), 10, 2)
horizontal_component(Z, M)
}
