% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{inverse_proc_transform}
\alias{inverse_proc_transform}
\title{Inverse of a fitted similarity / orthogonal transform.}
\usage{
inverse_proc_transform(tr)
}
\arguments{
\item{tr}{A \code{proc_fitted_transform}.}
}
\value{
A \code{proc_fitted_transform}.
}
\description{
Inverse of a fitted similarity / orthogonal transform.
}
