% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{overlap_graph}
\alias{overlap_graph}
\title{Overlap graph of configurations.}
\usage{
overlap_graph(data, min_overlap = 1L)
}
\arguments{
\item{data}{A \code{proc_data} object.}

\item{min_overlap}{Minimum shared observed entities for an edge.}
}
\value{
A list with adjacency metadata and connected components.
}
\description{
Overlap graph of configurations.
}
