% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{predict.gpa_fit}
\alias{predict.gpa_fit}
\title{Align new data to a fitted consensus.}
\usage{
\method{predict}{gpa_fit}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{gpa_fit}.}

\item{newdata}{A matrix or \code{proc_data} with one view.}

\item{...}{Unused.}
}
\description{
Align new data to a fitted consensus.
}
