% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{principal_angles}
\alias{principal_angles}
\title{Principal angles between two subspaces.}
\usage{
principal_angles(U, V)
}
\arguments{
\item{U, V}{Matrices whose columns span the subspaces.}
}
\value{
Numeric vector of principal angles in \eqn{[0, \pi/2]}.
}
\description{
Principal angles between two subspaces.
}
\examples{
principal_angles(diag(2), matrix(c(0, 1, 1, 0), 2, 2))
}
