% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{proc_from_array}
\alias{proc_from_array}
\title{Domain adapter: a 3-way entity-by-dimension-by-view array.}
\usage{
proc_from_array(x, view_names = NULL)
}
\arguments{
\item{x}{An array with dimensions \eqn{(n, d, K)} or a list of matrices.}

\item{view_names}{Optional view names.}
}
\description{
Domain adapter: a 3-way entity-by-dimension-by-view array.
}
