% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{proc_identity}
\alias{proc_identity}
\title{Identity fitted transform.}
\usage{
proc_identity(spec, d)
}
\arguments{
\item{spec}{A transform specification.}

\item{d}{Target dimension.}
}
\description{
Identity fitted transform.
}
