% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{proc_orthogonal}
\alias{proc_orthogonal}
\title{Orthogonal transformation family.}
\usage{
proc_orthogonal(group = c("O", "SO"))
}
\arguments{
\item{group}{\code{"O"} (reflections allowed) or \code{"SO"} (proper rotations).}
}
\value{
A \code{proc_transform_spec} object.
}
\description{
Orthogonal transformation family.
}
