% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{proc_signed_permutation}
\alias{proc_signed_permutation}
\title{Signed-permutation transformation family.}
\usage{
proc_signed_permutation()
}
\value{
A \code{proc_transform_spec} object.
}
\description{
Signed-permutation transformation family.
}
