% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{proc_similarity}
\alias{proc_similarity}
\title{Similarity transformation family.}
\usage{
proc_similarity(
  group = c("SO", "O"),
  translation = TRUE,
  scaling = c("isotropic", "none")
)
}
\arguments{
\item{group}{Orthogonal subgroup for the linear part.}

\item{translation}{Logical; estimate a translation.}

\item{scaling}{\code{"none"} or \code{"isotropic"}.}
}
\value{
A \code{proc_transform_spec} object.
}
\description{
Similarity transformation family.
}
