% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{transformations}
\alias{transformations}
\title{Fitted transforms.}
\usage{
transformations(object, ...)
}
\arguments{
\item{object}{A \code{gpa_fit}.}

\item{...}{Unused.}
}
\description{
Fitted transforms.
}
