% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{vertical_component}
\alias{vertical_component}
\title{Vertical / orbit part of a consensus perturbation.}
\usage{
vertical_component(Z, M, spec = proc_orthogonal("O"))
}
\arguments{
\item{Z}{Perturbation matrix.}

\item{M}{Consensus.}

\item{spec}{Transform family that defines the orbit.}
}
\value{
A matrix the same shape as \code{Z}.
}
\description{
Vertical / orbit part of a consensus perturbation.
}
\examples{
M <- scale(matrix(rnorm(20), 10, 2), scale = FALSE)
Z <- matrix(rnorm(20), 10, 2)
vertical_component(Z, M)
}
