% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{compile_proc_problem}
\alias{compile_proc_problem}
\title{Compile a problem into planner input.}
\usage{
compile_proc_problem(
  data,
  transform = proc_similarity(),
  gauge = proc_gauge(scale = "none"),
  metric = proc_metric(),
  loss = proc_squared_l2()
)
}
\arguments{
\item{data}{\code{proc_data}.}

\item{transform}{Transform spec.}

\item{gauge}{Gauge spec.}

\item{metric}{Metric.}

\item{loss}{Loss.}
}
\description{
Compile a problem into planner input.
}
