% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauges-metrics-losses.R
\name{proc_metric}
\alias{proc_metric}
\alias{proc_weights}
\title{Superimposition metric / weight channels.}
\usage{
proc_metric(
  configuration = NULL,
  landmark = NULL,
  cell = NULL,
  precision = NULL
)

proc_weights(
  configuration = NULL,
  landmark = NULL,
  cell = NULL,
  precision = NULL
)
}
\arguments{
\item{configuration}{Configuration weights \(\alpha_i\).}

\item{landmark}{Landmark / entity weights.}

\item{cell}{Coordinate-specific weights.}

\item{precision}{Precision operator \(Q\).}
}
\description{
Superimposition metric / weight channels.
}
